+++ packages/data-model/src/misc/AcDbRenderingCache.ts
@@ -210,7 +210,10 @@ export class AcDbRenderingCache {
         block = renderer.group(results)
         if (block && cache) {
           // If one block is one transient block whose name starts with '*U', don't cache it
-          if (AcDbBlockTableRecord.name && !AcDbBlockTableRecord.name.startsWith('*U')) {
+          if (
+            AcDbBlockTableRecord.name &&
+            !AcDbBlockTableRecord.name.startsWith('*U')
+          ) {
             this.set(key, block)
           }
         }